LPS-135286 move, Hugo SF, type: git ls-files **PortletSharedSearchContributor.java
